Optimize the parsing function and make keypad look 'better'
Changed the way the keypad display is read by the function. Respects
order of operation rules (not like it didn't), but it cleaned up the
code to a hell of a lot less lines of code. It reads more easily, too.
The display of the keypad was changed so the 'non-numbers' are colored
differently from the number keys. Still has a very clean look.